# fastapi/测试脚本/v2测试脚本/test/导出接口文档脚本/fastapi_complete_api_spec.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "用户系统API",
    "description": "模块化的用户管理系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v2/file_upload/validate": {
      "post": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Validate File Only",
        "description": "仅验证文件，不保存",
        "operationId": "validate_file_only_api_v2_file_upload_validate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_file_only_api_v2_file_upload_validate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/file_upload/upload": {
      "post": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Upload File",
        "description": "上传文件",
        "operationId": "upload_file_api_v2_file_upload_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v2_file_upload_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/file_upload/uploads": {
      "get": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Get Upload History",
        "description": "获取用户的文件上传历史",
        "operationId": "get_upload_history_api_v2_file_upload_uploads_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "状态筛选",
              "title": "Status Filter"
            },
            "description": "状态筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/file_upload/uploads/{upload_id}": {
      "get": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Get Upload Detail",
        "description": "获取指定上传记录的详细信息",
        "operationId": "get_upload_detail_api_v2_file_upload_uploads__upload_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Delete Upload Record",
        "description": "删除上传记录",
        "operationId": "delete_upload_record_api_v2_file_upload_uploads__upload_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/file_upload/create-document": {
      "post": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Create Document From Upload",
        "description": "从上传的文件创建文档",
        "operationId": "create_document_from_upload_api_v2_file_upload_create_document_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentFromUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Document From Upload Api V2 File Upload Create Document Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/file_upload/stats": {
      "get": {
        "tags": [
          "v2-file_upload"
        ],
        "summary": "Get Upload Stats",
        "description": "获取用户的上传统计信息",
        "operationId": "get_upload_stats_api_v2_file_upload_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_v2_file_upload_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "要上传的文件"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v2_file_upload_upload_post"
      },
      "Body_validate_file_only_api_v2_file_upload_validate_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "要验证的文件"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_validate_file_only_api_v2_file_upload_validate_post"
      },
...
      "CreateDocumentFromUploadRequest": {
        "properties": {
          "upload_id": {
            "type": "integer",
            "title": "Upload Id",
            "description": "上传记录ID"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "文档标题"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "文档摘要"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id",
            "description": "文件夹ID"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "title"
        ],
        "title": "CreateDocumentFromUploadRequest",
        "description": "从上传文件创建文档的请求模型"
      },
...
      "FileListResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/UploadRecordResponse"
            },
            "type": "array",
            "title": "Files"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "files",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "FileListResponse",
        "description": "文件列表响应模型"
      },
...
      "FileUploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "上传是否成功"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "upload_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Id",
            "description": "上传记录ID"
          },
          "file_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Info",
            "description": "文件信息"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "FileUploadResponse",
        "description": "文件上传响应模型"
      },
      "FileValidationResponse": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid",
            "description": "文件是否有效"
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "检测到的文件类型"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "文件大小(字节)"
          },
          "validation_details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Validation Details",
            "description": "验证详情"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "错误信息"
          }
        },
        "type": "object",
        "required": [
          "is_valid",
          "file_type",
          "file_size",
          "validation_details"
        ],
        "title": "FileValidationResponse",
        "description": "文件验证响应模型"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadRecordResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "validation_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "file_size",
          "file_type",
          "status",
          "validation_message",
          "created_at",
          "updated_at"
        ],
        "title": "UploadRecordResponse",
        "description": "上传记录响应模型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}